zng_wgt_grid::cell

Function at

pub fn at(
    child: impl UiNode,
    column_row: impl IntoVar<(usize, usize)>,
) -> impl UiNode

P Cell column and row indexes.

If set to AT_AUTO the cell is positioned based on the logical index.

Is (0, 0) by default.

This property sets the INFO_ID.

See also the column or row properties to bind each index individually.