Function zng_wgt_grid::cell::column

pub fn column(child: impl UiNode, col: impl IntoVar<usize>) -> impl UiNode

P Cell column index.

If set to usize::MAX the cell is positioned based on the logical index.

Is 0 by default.

This property sets the INFO_ID.

See also the at property to bind both indexes at the same time.