zng_wgt_grid::cell

Function row_span

pub fn row_span(child: impl IntoUiNode, span: impl IntoVar<usize>) -> UiNode

P Cell row span.

Number of cells this one spans over vertically, starting from the row index and spanning down.

Is 1 by default, the index is clamped between 1..max where max is the maximum number of valid rows down from the cell column index.

Note that the cell will not influence the row height if it spans over multiple rows.

This property sets the INFO_ID.

See also the span property to bind both spans at the same time.