zng_wgt_grid
Module column

Column widget and properties.

Structs

Column

W Grid column definition.

Functions

get_index

P Get the column index.

get_index_len

P Get the column index and number of columns.

get_rev_index

P Get the column index, starting from the last column at 0.

is_even

P If the column index is even.

is_first

P If the column is the first.

is_last

P If the column is the last.

is_odd

P If the column index is odd.