Function zng_wgt_grid::auto_grow_mode

pub fn auto_grow_mode(mode: impl IntoVar<AutoGrowMode>)

c Defines the direction the grid auto-grows and the maximum inclusive index that can be covered by auto-generated columns or rows. If a cell is outside this index and is not covered by predefined columns or rows a new one is auto generated for it, but if the cell is also outside this max it is collapsed.

Is `AutoGrowMode::rows() by default.

Capture-Only

This property is capture-only, it only defines a property signature, it does not implement any behavior by itself. Widgets can capture and implement this property as part of their intrinsics, otherwise it will have no effect if set on a widget that does not implement it.