zng_wgt_grid
Struct Cell

pub struct Cell(/* private fields */);

W Grid cell container.

This widget defines properties that position and size widgets in a Grid!.

See the Grid::cells property for more details.

Implementations

impl Cell

pub fn widget_new() -> Self

Start building a new instance.

pub fn widget_type() -> WidgetType

Gets the widget type info.

impl Cell

pub fn column(&self, col: impl IntoVar<usize>)

P Cell column index.

If set to usize::MAX the cell is positioned based on the logical index.

Is 0 by default.

This property sets the INFO_ID.

See also the at property to bind both indexes at the same time.

impl Cell

pub fn row(&self, row: impl IntoVar<usize>)

P Cell row index.

If set to usize::MAX the cell is positioned based on the logical index.

Is 0 by default.

This property sets the INFO_ID.

See also the at property to bind both indexes at the same time.

impl Cell

pub fn at(&self, column_row: impl IntoVar<(usize, usize)>)

P Cell column and row indexes.

If set to AT_AUTO the cell is positioned based on the logical index.

Is (0, 0) by default.

This property sets the INFO_ID.

See also the column or row properties to bind each index individually.

impl Cell

pub fn column_span(&self, span: impl IntoVar<usize>)

P Cell column span.

Number of cells this one spans over horizontally, starting from the column index and spanning to the right.

Is 1 by default, the index is clamped between 1..max where max is the maximum number of valid columns to the right of the cell column index.

Note that the cell will not influence the column width if it spans over multiple columns.

This property sets the INFO_ID.

See also the span property to bind both spans at the same time.

impl Cell

pub fn row_span(&self, span: impl IntoVar<usize>)

P Cell row span.

Number of cells this one spans over vertically, starting from the row index and spanning down.

Is 1 by default, the index is clamped between 1..max where max is the maximum number of valid rows down from the cell column index.

Note that the cell will not influence the row height if it spans over multiple rows.

This property sets the INFO_ID.

See also the span property to bind both spans at the same time.

impl Cell

pub fn span(&self, span: impl IntoVar<(usize, usize)>)

P Cell column and row span.

Is (1, 1) by default.

This property sets the INFO_ID.

See also the column_span or row_span properties to bind each index individually.