Crate zng_wgt_grid

This crate is part of the zng project.

Grid widgets, properties and nodes.

Modules

cell

Cell widget and properties.

column

Column widget and properties.

row

Row widget and properties.

Macros

Cell

Column

Row

Structs

AutoGrowFnArgs

Arguments for grid::auto_grow_fn.

Cell

W Grid cell container.

Column

W Grid column definition.

Grid

W Grid layout with cells of variable sizes.

Row

W Grid row definition.

Enums

AutoGrowMode

Grid auto-grow direction.

Functions

auto_grow_fn

c Widget function used when new rows or columns are needed to cover a cell placement.

auto_grow_mode

c Defines the direction the grid auto-grows and the maximum inclusive index that can be covered by auto-generated columns or rows. If a cell is outside this index and is not covered by predefined columns or rows a new one is auto generated for it, but if the cell is also outside this max it is collapsed.

cells

c Cell widget items.

columns

c Column definitions.

node

Grid node.

rows

c Row definitions.

spacing

c Space in-between cells.